vuuno4kse: Fix multiple recipes
NOTE: Resolving any missing task queue dependencies
NOTE: Multiple providers are available for runtime vuplus-bluetooth-util (vuplus-bluetooth-util-vuuno4k, vuplus-bluetooth-util-vuuno4kse)
Consider defining a PREFERRED_RPROVIDER entry to match vuplus-bluetooth-util

For some strange reason, PROVIDES and RPROVIDES can't be in the .inc file, but must be in the MACHINE specific config file.
Due to this, it needs to be done for each box as well.